rename from `string-undump` to `string_undump` due to naming convention
