+++ README.md
@@ -6,21 +6,21 @@ See [Feature #12275](https://bugs.ruby-lang.org/issues/12275) for details.
 ## Usage
 
 ```ruby
-require 'string-undump'
+require 'string_undump'
 
 puts '\u3059\u3054\u30FC\u3044'.undump #=> すごーい
 puts '\xE3\x81\x9F\xE3\x81\xAE\xE3\x81\x97\xE3\x83\xBC'.undump #=> たのしー
 puts '\u{1F43E}'.undump #=> 🐾
 ```
 
-See [testcase](https://github.com/tadd/string-undump/blob/master/test/test_undump.rb) also.
+See [testcase](https://github.com/tadd/string_undump/blob/master/test/test_undump.rb) also.
 
 ## Installation
 
 Add this line to your application's Gemfile:
 
 ```ruby
-gem 'string-undump'
+gem 'string_undump'
 ```
 
 And then execute:
@@ -29,7 +29,7 @@ And then execute:
 
 Or install it yourself as:
 
-    $ gem install string-undump
+    $ gem install string_undump
 
 ## Supported Ruby versions
 
@@ -37,7 +37,7 @@ Or install it yourself as:
 
 ## Contributing
 
-Bug reports and pull requests are welcome on GitHub at https://github.com/tadd/string-undump.
+Bug reports and pull requests are welcome on GitHub at https://github.com/tadd/string_undump.
 
 ## License
