+++ Rakefile
@@ -6,8 +6,8 @@ task :build => :compile
 
 Rake::TestTask.new {|t| t.libs << 'test' }
 
-Rake::ExtensionTask.new("string-undump") do |ext|
-  ext.lib_dir = "lib/string-undump"
+Rake::ExtensionTask.new("string_undump") do |ext|
+  ext.lib_dir = "lib/string_undump"
 end
 
 task :default => :test
