--- ext/string-undump/extconf.rb
@@ -1,3 +1,0 @@
-require 'mkmf'
-
-create_makefile("string-undump/string-undump")
