+++ ext/string_undump/extconf.rb
@@ -1,0 +1,3 @@
+require 'mkmf'
+
+create_makefile("string_undump/string_undump")
