+++ test/test_undump.rb
@@ -1,6 +1,6 @@
 require 'test-unit'
 
-require_relative '../lib/string-undump'
+require_relative '../lib/string_undump'
 
 class TestUndump < Test::Unit::TestCase
   def test_undump_badly
